pyproject.toml
[tool.poetry]
name = "stock_prediction"
version = "0.1.0"
description = ""
authors = ["fraser-brownn <fraserdbrown@yahoo.co.uk>"]

[tool.poetry.dependencies]
python = "^3.9"
psycopg2-binary = "^2.9.6"
pandas = "^2.0.3"
yfinance = "^0.2.22"
numpy = "^1.25.0"
scikit-learn = "^1.3.0"
matplotlib = "^3.7.2"
ipykernel = "^6.24.0"
fastapi = "^0.100.1"
uvicorn = "^0.23.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
